added qte tranche

+++ database/migrations/2021_03_20_084634_create_stock_poids_pcs_table.php
@@ -16,13 +16,13 @@ class CreateStockPoidsPcsTable extends Migration
         Schema::create('stock_poids_pcs', function (Blueprint $table) {
             $table->id();
             $table->integer('qte');
-            $table->double('prix_achat');
+            $table->decimal('prix_achat');
             $table->string('code');
             $table->double('poids');
-            $table->double('cr');
-            $table->double('prix_n');
-            $table->double('prix_f');
-            $table->double('prix_p');
+            $table->decimal('cr');
+            $table->decimal('prix_n');
+            $table->decimal('prix_f');
+            $table->decimal('prix_p');
             $table->decimal('pas');
 
             $table->unsignedBigInteger('unite_id');
